x11/linux: Support for XRender and Scaling
This adds support for scaling the window of the emulator. On 4K screens,
GSport window is really tiny and unreadable. This implements another
stage to the rendering using the XRender extension with a scaling
transform.

This appears to work with or without shared memory

Also, modifies the Makefile to easily switch scaling on or off --
technically if could be made configurable at runtime too, for the moment
the constant is passed as a #define

Signed-off-by: Michel Pollet <buserror@gmail.com>